+++ pypiserver/_app.py
@@ -5,7 +5,6 @@ import re
 import xml.dom.minidom
 import xmlrpc.client as xmlrpclib
 import zipfile
-
 from collections import namedtuple
 from io import BytesIO
 from urllib.parse import urljoin, urlparse
@@ -13,7 +12,6 @@ from urllib.parse import urljoin, urlparse
 from pypiserver.config import RunConfig
 from . import __version__
 from . import core
-from .backend import Backend
 from .bottle import (
     static_file,
     redirect,
@@ -23,13 +21,10 @@ from .bottle import (
     Bottle,
     template,
 )
-
-
 from .pkg_helpers import guess_pkgname_and_version, normalize_pkgname_for_url
 
 log = logging.getLogger(__name__)
 config: RunConfig
-backend: Backend
 app = Bottle()
 
 
@@ -101,7 +96,7 @@ def root():
         msg,
         URL=request.url.rstrip("/") + "/",
         VERSION=__version__,
-        NUMPKGS=backend.package_count(),
+        NUMPKGS=config.backend.package_count(),
         PACKAGES=fp.rstrip("/") + "/packages/",
         SIMPLE=fp.rstrip("/") + "/simple/",
     )
@@ -135,11 +130,11 @@ def remove_pkg():
         msg = f"Missing 'name'/'version' fields: name={name}, version={version}"
         raise HTTPError(400, msg)
 
-    pkgs = list(backend.find_version(name, version))
+    pkgs = list(config.backend.find_version(name, version))
     if not pkgs:
         raise HTTPError(404, f"{name} ({version}) not found")
     for pkg in pkgs:
-        backend.remove_package(pkg)
+        config.backend.remove_package(pkg)
 
 
 Upload = namedtuple("Upload", "pkg sig")
@@ -169,7 +164,7 @@ def file_upload():
         ):
             raise HTTPError(400, f"Bad filename: {uf.raw_filename}")
 
-        if not config.overwrite and backend.exists(uf.raw_filename):
+        if not config.overwrite and config.backend.exists(uf.raw_filename):
             log.warning(
                 f"Cannot upload {uf.raw_filename!r} since it already exists! \n"
                 "  You may start server with `--overwrite` option. "
@@ -180,7 +175,7 @@ def file_upload():
                 "  You may start server with `--overwrite` option.",
             )
 
-        backend.add_package(uf.raw_filename, uf.file)
+        config.backend.add_package(uf.raw_filename, uf.file)
         if request.auth:
             user = request.auth[0]
         else:
@@ -237,7 +232,7 @@ def handle_rpc():
         )
         response = []
         ordering = 0
-        for p in backend.get_all_packages():
+        for p in config.backend.get_all_packages():
             if p.pkgname.count(value) > 0:
                 # We do not presently have any description/summary, returning
                 # version instead
@@ -258,7 +253,7 @@ def handle_rpc():
 @app.route("/simple/")
 @auth("list")
 def simpleindex():
-    links = sorted(backend.get_projects())
+    links = sorted(config.backend.get_projects())
     tmpl = """\
     <html>
         <head>
@@ -284,7 +279,7 @@ def simple(project):
         return redirect(f"/simple/{normalized}/", 301)
 
     packages = sorted(
-        backend.find_project_packages(project),
+        config.backend.find_project_packages(project),
         key=lambda x: (x.parsed_version, x.relfn),
     )
     if not packages:
@@ -323,11 +318,11 @@ def simple(project):
 def list_packages():
     fp = request.custom_fullpath
     packages = sorted(
-        backend.get_all_packages(),
+        config.backend.get_all_packages(),
         key=lambda x: (os.path.dirname(x.relfn), x.pkgname, x.parsed_version),
     )
 
-    links =(
+    links = (
         (pkg.relfn_unix, urljoin(fp, pkg.fname_and_hash)) for pkg in packages
     )
 
@@ -350,7 +345,7 @@ def list_packages():
 @app.route("/packages/:filename#.*#")
 @auth("download")
 def server_static(filename):
-    entries = backend.get_all_packages()
+    entries = config.backend.get_all_packages()
     for x in entries:
         f = x.relfn_unix
         if f == filename:
