+++ pypiserver/core.py
@@ -1,32 +1,15 @@
 #! /usr/bin/env python3
 """minimal PyPI like server for use with pip/easy_install"""
 
-import functools
 import mimetypes
 import typing as t
-from typing import Optional
 from urllib.parse import quote
 
-from .backend import (
-    Backend,
-    SimpleFileBackend,
-    PkgFile,
-    CachingFileBackend,
-    IBackend,
-)
-from .cache import ENABLE_CACHING, CacheManager
-
-backend: Optional[Backend] = None
+from pypiserver.pkg_helpers import normalize_pkgname, parse_version
 
 mimetypes.add_type("application/octet-stream", ".egg")
 mimetypes.add_type("application/octet-stream", ".whl")
 mimetypes.add_type("text/plain", ".asc")
-
-
-def get_file_backend(config) -> SimpleFileBackend:
-    if ENABLE_CACHING:
-        return CachingFileBackend(config, config.roots, CacheManager())
-    return SimpleFileBackend(config, config.roots)
 
 
 def get_bad_url_redirect_path(request, project):
@@ -40,51 +23,59 @@ def get_bad_url_redirect_path(request, project):
     return uri
 
 
-PkgFunc = t.TypeVar("PkgFunc", bound=t.Callable[..., t.Iterable[PkgFile]])
+class PkgFile:
+    __slots__ = [
+        "pkgname",  # The projects/package name with possible capitalization
+        "version",  # The package version as a string
+        "fn",  # The full file path
+        "root",  # An optional root directory of the file
+        "relfn",  # The file path relative to the root
+        "replaces",  # The previous version of the package (used by manage.py)
+        "pkgname_norm",  # The PEP503 normalized project name
+        "digest",  # The file digest in the form of <algo>=<hash>
+        "relfn_unix",  # The relative file path in unix notation
+        "parsed_version",  # The package version as a tuple of parts
+        "digester",  # a function that calculates the digest for the package
+    ]
+    digest: t.Optional[str]
+    digester: t.Optional[t.Callable[["PkgFile"], t.Optional[str]]]
+    parsed_version: tuple
+    relfn_unix: t.Optional[str]
 
+    def __init__(
+        self,
+        pkgname: str,
+        version: str,
+        fn: t.Optional[str] = None,
+        root: t.Optional[str] = None,
+        relfn: t.Optional[str] = None,
+        replaces: t.Optional["PkgFile"] = None,
+    ):
+        self.pkgname = pkgname
+        self.pkgname_norm = normalize_pkgname(pkgname)
+        self.version = version
+        self.parsed_version: tuple = parse_version(version)
+        self.fn = fn
+        self.root = root
+        self.relfn = relfn
+        self.relfn_unix = None if relfn is None else relfn.replace("\\", "/")
+        self.replaces = replaces
+        self.digest = None
 
-def with_digester(func: PkgFunc) -> PkgFunc:
-    @functools.wraps(func)
-    def add_digester_method(self, *args, **kwargs):
-        packages = func(self, *args, **kwargs)
-        for package in packages:
-            package.digester = self.backend.digest
-            yield package
+    def __repr__(self) -> str:
+        return "{}({})".format(
+            self.__class__.__name__,
+            ", ".join(
+                [
+                    f"{k}={getattr(self, k, 'AttributeError')!r}"
+                    for k in sorted(self.__slots__)
+                ]
+            ),
+        )
 
-    return add_digester_method
-
-
-class BackendProxy(IBackend):
-    def __init__(self, wraps: Backend):
-        self.backend = wraps
-
-    @with_digester
-    def get_all_packages(self) -> t.Iterable[PkgFile]:
-        return self.backend.get_all_packages()
-
-    @with_digester
-    def find_project_packages(self, project: str) -> t.Iterable[PkgFile]:
-        return self.backend.find_project_packages(project)
-
-    def find_version(self, name: str, version: str) -> t.Iterable[PkgFile]:
-        return self.backend.find_version(name, version)
-
-    def get_projects(self) -> t.Iterable[str]:
-        return self.backend.get_projects()
-
-    def exists(self, filename: str) -> bool:
-        assert "/" not in filename
-        return self.backend.exists(filename)
-
-    def package_count(self) -> int:
-        return self.backend.package_count()
-
-    def add_package(self, filename, fh: t.BinaryIO):
-        assert "/" not in filename
-        return self.backend.add_package(filename, fh)
-
-    def remove_package(self, pkg: PkgFile):
-        return self.backend.remove_package(pkg)
-
-    def digest(self, pkg: PkgFile) -> str:
-        return self.backend.digest(pkg)
+    @property
+    def fname_and_hash(self) -> str:
+        if self.digest is None and self.digester is not None:
+            self.digest = self.digester(self)
+        hashpart = f"#{self.digest}" if self.digest else ""
+        return self.relfn_unix + hashpart  # type: ignore
